style: vscode에 eslint 깔아서 linter에 맞게만 수정
코드 내용은 수정 안 함
