+++ backend/package.json
@@ -33,6 +33,7 @@
     "cookie-parser": "^1.4.6",
     "https-proxy-agent": "^7.0.0",
     "jsonwebtoken": "^9.0.0",
+    "nest": "^0.1.6",
     "otplib": "^12.0.1",
     "passport": "^0.6.0",
     "passport-42": "^1.2.6",
