+++ backend/src/app.controller.ts
@@ -6,7 +6,7 @@ export class AppController {
   constructor(private readonly appService: AppService) {}
 
   @Get()
-  hello(){
+  hello() {
     return 'Hello World!';
   }
 }
