+++ backend/src/auth/auth42/auth42.guard.ts
@@ -1,0 +1,5 @@
+import { Injectable } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
+
+@Injectable()
+export class AuthGuard42 extends AuthGuard('42') {}
