+++ backend/src/auth/interfaces/RequestWithUser.interface.ts
